api/requirements.txt
flask
sentence-transformers
faiss-cpu
langchain
ollama
torch
flask-cors
